BC-569 | Refactor of existing feature flag annotations to move on top of controller instead. Also moved classes to respected packages.
