--- src/main/java/uk/gov/justice/laa/amend/claim/config/features/RequiresFeatureFlag.java
@@ -1,13 +1,0 @@
-package uk.gov.justice.laa.amend.claim.config.features;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target({ElementType.TYPE, ElementType.METHOD})
-@Retention(RetentionPolicy.RUNTIME)
-public @interface RequiresFeatureFlag {
-
-  Feature[] value();
-}
