+++ src/main/java/uk/gov/justice/laa/amend/claim/controllers/amendments/ClientController.java
@@ -15,19 +15,19 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import uk.gov.justice.laa.amend.claim.annotations.HasRoleClaimAmendmentsCaseworker;
+import uk.gov.justice.laa.amend.claim.annotations.RequiresFeatureFlag;
 import uk.gov.justice.laa.amend.claim.config.features.Feature;
-import uk.gov.justice.laa.amend.claim.config.features.RequiresFeatureFlag;
 import uk.gov.justice.laa.amend.claim.forms.amendments.AmendmentForm;
 import uk.gov.justice.laa.amend.claim.viewmodels.claimclient.ClaimClientViewFactory;
 
 @Controller
+@RequiresFeatureFlag(Feature.CLAIM_AMENDMENT)
 @RequestMapping("/submissions/{submissionId}/claims/{claimId}/amendments")
 @HasRoleClaimAmendmentsCaseworker
 @RequiredArgsConstructor
 public class ClientController {
 
   @GetMapping("/client")
-  @RequiresFeatureFlag(Feature.CLAIM_AMENDMENT)
   public String viewClient(
       HttpSession session,
       Model model,
@@ -45,7 +45,6 @@ public class ClientController {
   }
 
   @GetMapping("/amend-client")
-  @RequiresFeatureFlag(Feature.CLAIM_AMENDMENT)
   public String viewAmendClient(
       HttpSession session,
       Model model,
@@ -63,7 +62,6 @@ public class ClientController {
   }
 
   @PostMapping("/amend-client")
-  @RequiresFeatureFlag(Feature.CLAIM_AMENDMENT)
   public String amendClient1(
       HttpSession session,
       @ModelAttribute("client1Form") AmendmentForm client1Form,
